platform dependant sensorLib (WIP)
